Rename to Quickclump
Rename, again. This time, hopefully, for the last time.  The previous name
Dendroclump was too close to the original DENDROFIND, so I felt it may have
been confusing, or even inappropriate.  Name Quickclump suggests that the
code is fast, it also seems noone uses the name yet, and it can be abbreviated
as qc, which seems not in used either.